requirements/requirements.txt
numpy
pyyaml
fire
xraylib
matplotlib